+++ pom.xml
@@ -1,7 +1,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>dropboxapi</groupId>
-  <artifactId>DropboxAPI</artifactId>
+  <artifactId>easydropboxapi</artifactId>
   <version>1.0.0</version>
   <build>
     <sourceDirectory>src</sourceDirectory>
@@ -24,6 +23,7 @@
   		<version>1.7.2</version>
   	</dependency>
   </dependencies>
-  <name>DropBox API</name>
-  <url>https://www.dropbox.com/developers/reference/devguide</url>
+  <url>https://github.com/ITSStartup/easyDropboxAPI</url>
+  <name>easyDropboxAPI</name>
+  <groupId>com.its.easydropbox.api</groupId>
 </project>
