+++ src/com/its/easydropbox/api/interfaces/EasyDropBox.java
@@ -1,0 +1,50 @@
+package com.its.easydropbox.api.interfaces;
+
+import java.util.List;
+
+import com.dropbox.core.DbxClient;
+import com.dropbox.core.DbxEntry;
+import com.dropbox.core.DbxEntry.File;
+import com.dropbox.core.DbxEntry.WithChildren;
+import com.dropbox.core.DbxException;
+
+public interface EasyDropBox {
+	/**
+	 * 
+	 * @return all files for dropboxPath
+	 * @throws DbxException
+	 */
+	List<DbxEntry> listFiles() throws DbxException;
+
+	/**
+	 * 
+	 * @return a File from dropbox
+	 */
+	File getFile();
+	/**
+	 * 
+	 * @param token generated by dropbox api
+	 * @return  files/folders for dropboxpath specified
+	 * @throws DbxException 
+	 */
+	WithChildren getListing() throws DbxException;
+
+	/**
+	 * 
+	 * @param basePath inform the path of the file 
+	 * @param query  what you wish search 
+	 * @return 
+	 * @throws DbxException 
+	 */
+	List<DbxEntry> searchFile(String basePath, String query) throws DbxException;
+
+	DbxClient getClient();
+	
+	/**
+	 * 
+	 * @param newpath must be absolute path 
+	 * e.g: /nameOfFolder
+	 */
+	void changeDropboxPath(String newpath);
+
+}
